This program finds out the factorial of large numbers
This program finds out the factorial of large numbers

For finding out the factorial of large numbers simple factorial program
using integer or long wont help.
So here is a java program which finds out the factorials of fairly large
numbers with the help of custom made linked list.

The program is written in java.

Cheers!!!